Bug fix in cases that rf > # distinct nodes
In cases that the replication factor is greater than the number of
distinct nodes in the ring, fixReplicaOwners breaks from its internal
(per virtual node) loop when it has cycled over the whole ring, leaving
an empty string in some slots for every key of the replicaOwners map.
This is now fixed by properly setting the length of that slice to be
lower than its allocated cap, deliberately leaving a leak there, which
is properly documented along with a possible mitigation solution, which
does not interest me for now.